fixed warning errors
