+++ utils/browserDetection.ts
@@ -1,0 +1,10 @@
+export const getiOSVersion = () => {
+  const userAgent = window.navigator.userAgent;
+  const match = userAgent.match(/OS (\d+)_/);
+  return match ? parseInt(match[1], 10) : null;
+};
+
+export const isSafari = () => {
+  const ua = navigator.userAgent.toLowerCase();
+  return ua.includes('safari') && !ua.includes('chrome');
+};
